+++ src/components/yoo-illustration/yoo-illustration.types.ts
@@ -1,5 +1,7 @@
+import { ILLUSTRATIONS } from './yoo-illustration-base';
+
 export type IllustrationProps = {
-  name: string;
+  name: keyof typeof ILLUSTRATIONS;
   width: number;
   height: number;
 };