+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Todas as mudanças notáveis deste projeto serão documentadas neste arquivo.
 
 Este projeto segue o [Keep a Changelog](https://keepachangelog.com/pt-BR/1.0.0/) e [Semantic Versioning](https://semver.org/lang/pt-BR/).
 
+## [1.2.0]
+
+- **Adicionado**
+  - Acompanhamento de Changelog pelo storybook.
+
+- **Corrigido**
+  - Corrigido name da prop no storybook do componente Yoo-Illustration.
+
 ## [1.1.0]
 
 - **Adicionado**